Refactor Booking and Notebook classes
Refactored the `Reserve` method in the `Booking` class to accept a `Notebook` object and a new `PricingService` parameter, removing the `notebookId` parameter. Updated the `Notebook` class by replacing the `Availability` property with a `LastBookedOnUtc` property, which tracks the last booking time. Adjusted the `Notebook` constructor to initialize `LastBookedOnUtc` and made this property `internal set` for controlled modification.